requirements.txt
fastapi
uvicorn
torch
transformers
google-generativeai
python-dotenv
chromadb
langchain
langchain-community
sentence-transformers
langchain-google-genai